requirements.txt
pyTelegramBotAPI==4.14.0
Pillow==10.1.0
cryptography==41.0.7
PyMuPDF==1.23.8
python-docx==1.1.0
pdf2docx==0.5.6
docx2pdf==0.1.8
reportlab==4.0.7
pandas==2.1.4
openpyxl==3.1.2
python-pptx==0.6.23